+++ source/extensions/omni.isaac.sensor/docs/CHANGELOG.md
@@ -4,9 +4,12 @@
 ### Added
 - Added support for ros and usd camera axes in get_world_pose, get_local_pose, set_world_pose, set_local_pose.
 - Added Effort Sensor
+- supported_annotators property
 ### Fixed
 - Fixed divisible by zero error in IMU linear interpolation
 - Removed reading pairs from the IMU sensor to use the buffer directly
+### Changed
+- store and destroy internaly created renderproduct
 
 ## [7.1.0] - 2023-08-04
 ### Added