+++ source/extensions/omni.isaac.sensor/docs/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [7.2.1] - 2023-08-10
+### Added 
+- docstrings to Camera class for functions adding annotators 
+
 ## [7.2.0] - 2023-08-08
 ### Added
 - Added support for ros and usd camera axes in get_world_pose, get_local_pose, set_world_pose, set_local_pose.